Only rename AccessLogValve file with renameOnrotate=true if target file does not already exist.
We don't want to overwrite existing files. Prefer
not to rotate.

Typical case would be several ALV instances configured to
use the same file name. Only the first one should rename
on rotate.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@1416529 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/catalina/valves/AccessLogValve.java
@@ -1086,13 +1086,17 @@ public class AccessLogValve extends ValveBase implements AccessLog {
         writer.close();
         if (rename && renameOnRotate) {
             File newLogFile = getLogFile(true);
-            try {
-                if (!currentLogFile.renameTo(newLogFile)) {
-                    log.error(sm.getString("accessLogValve.renameFail", currentLogFile, newLogFile));
+            if (!newLogFile.exists()) {
+                try {
+                    if (!currentLogFile.renameTo(newLogFile)) {
+                        log.error(sm.getString("accessLogValve.renameFail", currentLogFile, newLogFile));
+                    }
+                } catch (Throwable e) {
+                    ExceptionUtils.handleThrowable(e);
+                    log.error(sm.getString("accessLogValve.renameFail", currentLogFile, newLogFile), e);
                 }
-            } catch (Throwable e) {
-                ExceptionUtils.handleThrowable(e);
-                log.error(sm.getString("accessLogValve.renameFail", currentLogFile, newLogFile), e);
+            } else {
+                log.error(sm.getString("accessLogValve.alreadyExists", currentLogFile, newLogFile));
             }
         }
         writer = null;
